Move Chado Test Trait into the Trait namespace.
