+++ tripal_chado/tests/src/Functional/ChadoTestBrowserBase.php
@@ -2,6 +2,7 @@
 namespace Drupal\Tests\tripal_chado\Functional;
 
 use Drupal\Tests\tripal\Functional\TripalTestBrowserBase;
+use Drupal\Tests\tripal_chado\Traits\ChadoTestTrait;
 use Drupal\tripal\TripalDBX\TripalDbx;
 use Drupal\tripal_chado\Database\ChadoConnection;
