+++ tripal_chado/tests/src/Kernel/ChadoTestKernelBase.php
@@ -4,7 +4,7 @@ namespace Drupal\Tests\tripal_chado\Kernel;
 use Drupal\Tests\tripal\Kernel\TripalTestKernelBase;
 use Drupal\tripal\TripalDBX\TripalDbx;
 use Drupal\tripal_chado\Database\ChadoConnection;
-use Drupal\Tests\tripal_chado\Functional\ChadoTestTrait;
+use Drupal\Tests\tripal_chado\Traits\ChadoTestTrait;
 
 /**
  * This is a base class for Chado tests.
