+++ tripal_chado/tests/src/Traits/ChadoTestTrait.php
@@ -1,5 +1,5 @@
 <?php
-namespace Drupal\Tests\tripal_chado\Functional;
+namespace Drupal\Tests\tripal_chado\Traits;
 
 use Drupal\tripal\TripalDBX\TripalDbx;
 use Drupal\tripal_chado\Database\ChadoConnection;
